фикс правок
