+++ src/main/kotlin/nsu/fit/qyoga/app/questionnaires/createQuestionnaires/QuestionnairesResultsController.kt
@@ -92,9 +92,6 @@ class QuestionnairesResultsController(
     }
 
     fun setQuestionnaireInSession(questionnaire: CreateQuestionnaireDto) {
-        httpSession.setAttribute(
-            questionnaireFieldName,
-            questionnaire
-        )
+        httpSession.setAttribute(questionnaireFieldName, questionnaire)
     }
 }
